ListBox Study
In this project, I used listbox, radio button and checkbox elements that allow users to make choices. These elements are placed to facilitate user interaction and to offer different options.

Used Components and Functions:
Listbox: Listbox component allows multiple options to be displayed and allows the user to select one or more options. For example, it was used to allow users to choose between various categories. The user can send their information by selecting one or more options.

Radio Button: Radio button allows only one choice to be made among the options in the same group. In this project, it was used in cases where the user needs to choose between certain categories. For example, when users need to choose a payment method, I preferred this component because they only need to mark one option.

Checkbox: Checkbox component allows multiple options to be selected. In the project, it was used in cases where users need to determine their interests or select certain features. For example, I added checkboxes so that users can choose different notification options (e-mail, SMS, etc.).

How the Project Works:
When the user opens the project, an interface with specific options appears. While the listbox offers the user a wide range of options, more specific selections can be made with the radio button and checkbox. After the selection is made, the user can save their selections by pressing the "Send" button or save these selections to a database for later use.